# Optimize writing small strings