@JamesNK 2021-01-26
After:
```
|                           Method | BytesToWrite | encodedSize |       Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------------- |------------- |------------ |-----------:|----------:|----------:|------:|------:|------:|----------:|
|         WriteString_WriteContext |        10080 |           1 | 197.933 us | 3.4468 us | 3.2242 us |     - |     - |     - |         - |
|         WriteString_WriteContext |        10080 |           4 |  51.337 us | 0.2081 us | 0.1845 us |     - |     - |     - |         - |
| WriteNonAsciiString_WriteContext |        10080 |           4 |  63.139 us | 0.3039 us | 0.2843 us |     - |     - |     - |         - |
|         WriteString_WriteContext |        10080 |          10 |  21.122 us | 0.2113 us | 0.1977 us |     - |     - |     - |         - |
| WriteNonAsciiString_WriteContext |        10080 |          10 |  27.255 us | 0.5020 us | 0.4450 us |     - |     - |     - |         - |
```


Before:
```
|                           Method | BytesToWrite | encodedSize |       Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------------- |------------- |------------ |-----------:|----------:|----------:|------:|------:|------:|----------:|
|         WriteString_WriteContext |        10080 |           1 | 147.735 us | 0.7897 us | 0.7001 us |     - |     - |     - |         - |
|         WriteString_WriteContext |        10080 |           4 |  48.229 us | 0.7844 us | 0.6953 us |     - |     - |     - |       1 B |
| WriteNonAsciiString_WriteContext |        10080 |           4 | 117.600 us | 0.6227 us | 0.5520 us |     - |     - |     - |         - |
|         WriteString_WriteContext |        10080 |          10 |  25.140 us | 0.1824 us | 0.1706 us |     - |     - |     - |         - |
| WriteNonAsciiString_WriteContext |        10080 |          10 |  42.219 us | 0.6843 us | 0.6401 us |     - |     - |     - |         - |
```