@JamesNK 2021-01-27
> Is the new version (with the new WriteStringToBuffer method) faster, slower or the same as the previous version of this PR?

The same - https://github.com/protocolbuffers/protobuf/pull/8149#issuecomment-767901299